+++ src/csv.rs
@@ -5,4 +5,4 @@ pub(crate) mod reader;
 pub use reader::infer_schema;
 pub use reader::Config;
 pub use reader::FieldParser;
-pub use reader::Reader;
+pub use reader::{Reader, Record};